refactor: Move bet status in shared
